Commit 2 - Custom Classes for Gamemode and Character
I split us off from the default modes. There's a default level "CentralTestLevel" where we can test mechanics and all that. From the character, I removed the object grab system and my own changes, so we've got a blank slate to work with. Sprint already exists and works, but I'll tweak it and the walk speed to make something more like skating.